Begin separating header and implementation files.
