+++ src/applications/mapper/mapper-thread.hpp
@@ -25,101 +25,17 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+#pragma once
+
+#include <thread>
+#include <mutex>
 #include <random>
 
 #include "model/engine.hpp"
 #include "model/sparse-optimization-info.hpp"
+#include "search/search.hpp"
 
 extern bool gTerminate;
-
-enum class Betterness
-{
-  Better,
-  SlightlyBetter,
-  SlightlyWorse,
-  Worse
-};
-  
-static double Cost(const model::Topology::Stats& stats, const std::string metric)
-{
-  if (metric == "delay")
-  {
-    return static_cast<double>(stats.cycles);
-  }
-  else if (metric == "energy")
-  {
-    return stats.energy;
-  }
-  else if (metric == "last-level-accesses")
-  {
-    return stats.last_level_accesses;
-  }
-  else
-  {
-    assert(metric == "edp");
-    return (stats.energy * stats.cycles);
-  }
-}
-
-static Betterness IsBetterRecursive_(const model::Topology::Stats& candidate, const model::Topology::Stats& incumbent,
-                                     const std::vector<std::string>::const_iterator metric,
-                                     const std::vector<std::string>::const_iterator end)
-{
-  const double tolerance = 0.001;
-
-  double candidate_cost = Cost(candidate, *metric);
-  double incumbent_cost = Cost(incumbent, *metric);
-
-  double relative_improvement = incumbent_cost == 0 ? 1.0 :
-    (incumbent_cost - candidate_cost) / incumbent_cost;
-
-  if (fabs(relative_improvement) > tolerance)
-  {
-    // We have a clear winner.
-    if (relative_improvement > 0)
-      return Betterness::Better;
-    else
-      return Betterness::Worse;
-  }
-  else
-  {
-    // Within tolerance range, try to recurse.
-    if (std::next(metric) == end)
-    {
-      // Base case. NOTE! Equality is categorized as SlightlyWorse (prefers incumbent).
-      if (relative_improvement > 0)
-        return Betterness::SlightlyBetter;
-      else
-        return Betterness::SlightlyWorse;
-    }
-    else
-    {
-      // Recursive call.
-      Betterness lsm = IsBetterRecursive_(candidate, incumbent, std::next(metric), end);
-      if (lsm == Betterness::Better || lsm == Betterness::Worse)
-        return lsm;
-      // NOTE! Equality is categorized as SlightlyWorse (prefers incumbent).
-      else if (relative_improvement > 0)
-        return Betterness::SlightlyBetter;
-      else
-        return Betterness::SlightlyWorse;
-    }      
-  }
-}
-
-static inline bool IsBetter(const model::Topology::Stats& candidate, const model::Topology::Stats& incumbent,
-                            const std::vector<std::string>& metrics)
-{
-  Betterness b = IsBetterRecursive_(candidate, incumbent, metrics.begin(), metrics.end());
-  return (b == Betterness::Better || b == Betterness::SlightlyBetter);
-}
-
-static inline bool IsBetter(const model::Topology::Stats& candidate, const model::Topology::Stats& incumbent,
-                            const std::string& metric)
-{
-  std::vector<std::string> metrics = { metric };
-  return IsBetter(candidate, incumbent, metrics);
-}
 
 struct EvaluationResult
 {
@@ -127,19 +43,7 @@ struct EvaluationResult
   Mapping mapping;
   model::Topology::Stats stats;
 
-  bool UpdateIfBetter(const EvaluationResult& other, const std::vector<std::string>& metrics)
-  {
-    bool updated = false;
-    if (other.valid &&
-        (!valid || IsBetter(other.stats, stats, metrics)))
-    {
-      valid = true;
-      mapping = other.mapping;
-      stats = other.stats;
-      updated = true;
-    }
-    return updated;
-  }
+  bool UpdateIfBetter(const EvaluationResult& other, const std::vector<std::string>& metrics);
 };
 
 //--------------------------------------------//
@@ -152,17 +56,7 @@ enum class FailClass
   Capacity
 };
 
-std::map<FailClass, std::string> FailClassToString =
-{
-  { FailClass::Fanout, "Fanout" },
-  { FailClass::Capacity, "Capacity" }
-};
-
-std::ostream& operator << (std::ostream& out, const FailClass& fail_class)
-{
-  out << FailClassToString.at(fail_class);
-  return out;
-}
+std::ostream& operator << (std::ostream& out, const FailClass& fail_class);
 
 struct FailInfo
 {
@@ -186,65 +80,9 @@ class MapperThread
     std::default_random_engine generator;
     std::uniform_real_distribution<double> distribution;
 
-    Stats() :
-        distribution(0.0,1.0)
-    {
-    }
+    Stats();
 
-    void UpdateFails(FailClass fail_class, std::string fail_reason, unsigned level, const Mapping& mapping)
-    {
-        // Find the data corresponding to this fail class.
-        auto fail_bucket_it = fail_stats.find(fail_class);
-        if (fail_bucket_it == fail_stats.end())
-        {
-          // We've never seen this fail class before.
-          std::map<unsigned, FailInfo> fail_bucket;
-          fail_bucket[level] = { .count = 1, .mapping = mapping, .reason = fail_reason };
-          fail_stats[fail_class] = fail_bucket;
-        }
-        else
-        {
-          // We've seen this fail class, see if this level has
-          // failed in this class.
-          auto& fail_bucket = fail_bucket_it->second;
-          auto fail_info_it = fail_bucket.find(level);
-          if (fail_info_it == fail_bucket.end())
-          {
-            // No, this is the first time this level has failed in
-            // this fail class, create a new entry.
-            fail_bucket[level] = { .count = 1, .mapping = mapping, .reason = fail_reason };
-          }
-          else
-          {
-            // This level has already failed in this class,
-            // increment its count.
-            fail_info_it->second.count += 1;
- 
-            // p(x) = prob. that I switch to x when it arrives
-            // p(0) = 1
-
-            // P(x) = prob. that x is finally selected.
-            // 1/N = P(0) = p(0).(1-p(1)).(1-p(2))...(1-p(N-1))
-            // 1/N = P(1) =        (p(1)).(1-p(2))...(1-p(N-1))
-
-            // p(x).(1-p(x+1)) = p(x+1)
-            // ...
-            // => p(x+1) = p(x) / [1+p(x)]
-            // ...
-            // => p(x) = 1/(1+x)
-
-            // Compute the probability of switching (we've already computed count=x+1)
-            double prob = 1 / fail_info_it->second.count.convert_to<double>();
-
-            // Probabilistically update the mapping.
-            double roll = distribution(generator);
-            if (roll < prob)
-            {
-              fail_info_it->second.mapping = mapping;
-            }
-          }
-        }
-    }    
+    void UpdateFails(FailClass fail_class, std::string fail_reason, unsigned level, const Mapping& mapping);
   };
 
  private:
@@ -294,362 +132,14 @@ class MapperThread
     problem::Workload &workload,
     sparse::SparseOptimizationInfo* sparse_optimizations,
     EvaluationResult* best
-    ) :
-      thread_id_(thread_id),
-      search_(search),
-      mapspace_(mapspace),
-      mutex_(mutex),
-      search_size_(search_size),
-      timeout_(timeout),
-      victory_condition_(victory_condition),
-      sync_interval_(sync_interval),
-      log_stats_(log_stats),
-      log_suboptimal_(log_suboptimal),
-      log_stream_(log_stream),
-      live_status_(live_status),
-      diagnostics_on_(diagnostics_on),
-      penalize_consecutive_bypass_fails_(penalize_consecutive_bypass_fails),
-      optimization_metrics_(optimization_metrics),
-      arch_specs_(arch_specs),
-      workload_(workload),
-      sparse_optimizations_(sparse_optimizations),
-      best_(best),
-      thread_(),
-      stats_()
-  {
-  }
+    );
 
-  void Start()
-  {
-    // We can do this because std::thread is movable.
-    thread_ = std::thread(&MapperThread::Run, this);
-  }
+  void Start();
 
-  void Join()
-  {
-    thread_.join();
-  }
+  void Join();
 
-  const Stats& GetStats() const
-  {
-    return stats_;
-  }
+  const Stats& GetStats() const;
 
-  void Run()
-  {
-    uint128_t total_mappings = 0;
-    uint128_t valid_mappings = 0;
-    uint128_t invalid_mappings_mapcnstr = 0;
-    uint128_t invalid_mappings_eval = 0;
-    std::uint32_t mappings_since_last_best_update = 0;
-
-    const int ncurses_line_offset = 6;
+  void Run();
       
-    model::Engine engine;
-    engine.Spec(arch_specs_);
-
-    mapspace::ID prev_mapping_id;
-
-    // =================
-    // Main mapper loop.
-    // =================
-    while (true)
-    {
-      if (live_status_)
-      {
-        std::stringstream msg;
-
-        msg << std::setw(3) << thread_id_ << std::setw(11) << total_mappings
-            << std::setw(11) << (total_mappings - valid_mappings)  << std::setw(11) << valid_mappings
-            << std::setw(11) << invalid_mappings_mapcnstr + invalid_mappings_eval
-            << std::setw(11) << mappings_since_last_best_update;
-
-        if (valid_mappings > 0)
-        {
-          msg << std::setw(10) << std::fixed << std::setprecision(2) << (stats_.thread_best.stats.utilization * 100) << "%"
-              << std::setw(11) << std::fixed << std::setprecision(3) << stats_.thread_best.stats.energy /
-            stats_.thread_best.stats.algorithmic_computes;
-        }
-
-        mutex_->lock();
-        mvaddstr(thread_id_ + ncurses_line_offset, 0, msg.str().c_str());
-        refresh();
-        mutex_->unlock();
-      }
-
-      // Termination conditions.
-      bool terminate = false;
-
-      if (gTerminate)
-      {
-        mutex_->lock();
-        log_stream_ << "[" << std::setw(3) << thread_id_ << "] STATEMENT: "
-                    << "global termination flag activated, terminating search."
-                    << std::endl;
-        mutex_->unlock();
-        terminate = true;
-      }
-
-      if (search_size_ > 0 && valid_mappings == search_size_)
-      {
-        mutex_->lock();
-        log_stream_ << "[" << std::setw(3) << thread_id_ << "] STATEMENT: " << search_size_
-                    << " valid mappings found, terminating search."
-                    << std::endl;
-        mutex_->unlock();
-        terminate = true;
-      }
-
-      if (victory_condition_ > 0 && mappings_since_last_best_update == victory_condition_)
-      {
-        mutex_->lock();
-        log_stream_ << "[" << std::setw(3) << thread_id_ << "] STATEMENT: " << victory_condition_
-                    << " suboptimal mappings found since the last upgrade, terminating search."
-                    << std::endl;
-        mutex_->unlock();
-        terminate = true;
-      }
-        
-      if ((invalid_mappings_mapcnstr + invalid_mappings_eval) > 0 &&
-          (invalid_mappings_mapcnstr + invalid_mappings_eval) == timeout_)
-      {
-        mutex_->lock();
-        log_stream_ << "[" << std::setw(3) << thread_id_ << "] STATEMENT: " << timeout_
-                    << " invalid mappings (" << invalid_mappings_mapcnstr << " fanout, "
-                    << invalid_mappings_eval << " capacity) found since the last valid mapping, "
-                    << "terminating search." << std::endl;
-        mutex_->unlock();
-        terminate = true;
-      }
-
-      // Try to obtain the next mapping from the search algorithm.
-      mapspace::ID mapping_id;
-      if (!search_->Next(mapping_id))
-      {
-        mutex_->lock();
-        log_stream_ << "[" << std::setw(3) << thread_id_ << "] STATEMENT: "
-                    << "search algorithm is done, terminating search."
-                    << std::endl;        
-        mutex_->unlock();
-        terminate = true;
-      }
-
-      // Terminate.
-      if (terminate)
-      {
-        if (live_status_)
-        {
-          mutex_->lock();
-          mvaddstr(thread_id_ + ncurses_line_offset, 0, "-");
-          refresh();
-          mutex_->unlock();
-        }
-        break;
-      }
-
-      //
-      // Periodically sync thread_best with global best.
-      //
-      if (total_mappings != 0 && sync_interval_ > 0 && total_mappings % sync_interval_ == 0)
-      {
-        mutex_->lock();
-          
-        // Sync from global best to thread_best.
-        bool global_pulled = false;
-        if (best_->valid)
-        {
-          if (stats_.thread_best.UpdateIfBetter(*best_, optimization_metrics_))
-          {
-            global_pulled = true;
-          }
-        }
-
-        // Sync from thread_best to global best.
-        if (stats_.thread_best.valid && !global_pulled)
-        {
-          best_->UpdateIfBetter(stats_.thread_best, optimization_metrics_);
-        }
-          
-        mutex_->unlock();
-      }
-
-      //
-      // Check if the only change vs. the previous mapping was in the Bypass
-      // dimension. This is useful later.
-      //
-      bool only_bypass_changed = false;
-      if (total_mappings > 1)
-      {
-        bool match = true;
-        for (unsigned idim = 0; idim < unsigned(mapspace::Dimension::Num); idim++)
-        {
-          if (mapspace::Dimension(idim) != mapspace::Dimension::DatatypeBypass)
-            match &= (mapping_id[idim] == prev_mapping_id[idim]);
-        }
-        only_bypass_changed = match;
-      }
-      prev_mapping_id = mapping_id;
-
-      //
-      // Begin Mapping. We do this in several stages with increasing algorithmic
-      // complexity and attempt to bail out as quickly as possible at each stage.
-      //
-      bool success = true;
-
-      // Stage 1: Construct a mapping from the mapping ID. This step can fail
-      //          because the space of *legal* mappings isn't dense (unfortunately),
-      //          so a mapping ID may point to an illegal mapping.
-      Mapping mapping;
-
-      auto construction_status = mapspace_->ConstructMapping(mapping_id, &mapping, !diagnostics_on_);
-      success &= std::accumulate(construction_status.begin(), construction_status.end(), true,
-                                 [](bool cur, const mapspace::Status& status)
-                                 { return cur && status.success; });
-
-      total_mappings++;
-
-      if (!success)
-      {
-        invalid_mappings_mapcnstr++;
-        if (diagnostics_on_)
-        {
-          for (unsigned level = 0; level < construction_status.size(); level++)
-            if (!construction_status.at(level).success)
-              stats_.UpdateFails(FailClass::Fanout, construction_status.at(level).fail_reason, level, mapping);
-        }
-        search_->Report(search::Status::MappingConstructionFailure);
-        continue;
-      }
-
-      // Stage 2: (Re)Configure a hardware model to evaluate the mapping
-      //          on, and run some lightweight pre-checks that the
-      //          model can use to quickly reject a nest.
-      //engine.Spec(arch_specs_);
-      auto status_per_level = engine.PreEvaluationCheck(mapping, workload_, sparse_optimizations_, !diagnostics_on_);
-      success &= std::accumulate(status_per_level.begin(), status_per_level.end(), true,
-                                 [](bool cur, const model::EvalStatus& status)
-                                 { return cur && status.success; });
-
-      if (!success)
-      {
-        // Pre-evaluation failed.
-        // If the only change in this mapping vs. the previous mapping was in
-        // its dataspace bypass scheme, then we may not want to make this
-        // failure count towards the timeout termination trigger.
-        if (penalize_consecutive_bypass_fails_ || !only_bypass_changed)
-        {
-          invalid_mappings_eval++;
-        }
-
-        if (diagnostics_on_)
-        {
-          for (unsigned level = 0; level < status_per_level.size(); level++)
-            if (!status_per_level.at(level).success)
-              stats_.UpdateFails(FailClass::Capacity, status_per_level.at(level).fail_reason, level, mapping);
-        }
-        search_->Report(search::Status::EvalFailure);
-        continue;
-      }
-
-      // Stage 3: Heavyweight evaluation.
-      status_per_level = engine.Evaluate(mapping, workload_, sparse_optimizations_, !diagnostics_on_);
-      success &= std::accumulate(status_per_level.begin(), status_per_level.end(), true,
-                                 [](bool cur, const model::EvalStatus& status)
-                                 { return cur && status.success; });
-      if (!success)
-      {
-        // Evaluation failed.
-        // If the only change in this mapping vs. the previous mapping was in
-        // its dataspace bypass scheme, then we may not want to make this
-        // failure count towards the timeout termination trigger.
-        if (penalize_consecutive_bypass_fails_ || !only_bypass_changed)
-        {
-          invalid_mappings_eval++;
-        }
-
-        if (diagnostics_on_)
-        {
-          for (unsigned level = 0; level < status_per_level.size(); level++)
-            if (!status_per_level.at(level).success)
-              stats_.UpdateFails(FailClass::Capacity, status_per_level.at(level).fail_reason, level, mapping);
-        }
-        search_->Report(search::Status::EvalFailure);
-        continue;
-      }
-
-      // SUCCESS!!!
-      auto stats = engine.GetTopology().GetStats();
-      EvaluationResult result = { true, mapping, stats };
-
-      valid_mappings++;
-      if (log_stats_)
-      {
-        mutex_->lock();
-        log_stream_ << "[" << thread_id_ << "] INVALID " << total_mappings << " " << valid_mappings
-                    << " " << invalid_mappings_mapcnstr + invalid_mappings_eval << std::endl;
-        mutex_->unlock();
-      }        
-      invalid_mappings_mapcnstr = 0;
-      invalid_mappings_eval = 0;
-      search_->Report(search::Status::Success, Cost(stats, optimization_metrics_.at(0)));
-
-      if (log_suboptimal_)
-      {
-        mutex_->lock();
-        log_stream_ << "[" << std::setw(3) << thread_id_ << "]" 
-                    << " Utilization = " << std::setw(4) << std::fixed << std::setprecision(2) << stats.utilization 
-                    << " | pJ/Algorithmic-Compute = " << std::setw(8) << std::fixed << std::setprecision(3) << stats.energy / stats.algorithmic_computes
-                    << " | pJ/Compute = " << std::setw(8) << std::fixed << std::setprecision(3) << stats.energy / stats.actual_computes
-                    << " | " << mapping.PrintCompact()
-                    << std::endl;
-        mutex_->unlock();
-      }
-
-      // Is the new mapping "better" than the previous best mapping?
-      if (stats_.thread_best.UpdateIfBetter(result, optimization_metrics_))
-      {
-        if (log_stats_)
-        {
-          // FIXME: improvement only captures the primary stat.
-          double improvement = stats_.thread_best.valid ?
-            (Cost(stats_.thread_best.stats, optimization_metrics_.at(0)) - Cost(stats, optimization_metrics_.at(0))) /
-            Cost(stats_.thread_best.stats, optimization_metrics_.at(0)) : 1.0;
-          mutex_->lock();
-          log_stream_ << "[" << thread_id_ << "] UPDATE " << total_mappings << " " << valid_mappings
-                      << " " << mappings_since_last_best_update << " " << improvement << std::endl;
-          mutex_->unlock();
-        }
-        
-        if (!log_suboptimal_)
-        {
-          mutex_->lock();
-          log_stream_ << "[" << std::setw(3) << thread_id_ << "]"
-                      << " Utilization = " << std::setw(4) << std::fixed << std::setprecision(2) << stats.utilization
-                      << " | pJ/Algorithmic-Compute = " << std::setw(8) << std::fixed << std::setprecision(3)
-                      << stats.energy / stats.algorithmic_computes
-                      << " | pJ/Compute = " << std::setw(8) << std::fixed << std::setprecision(3)
-                      << stats.energy / stats.actual_computes
-                      << " | " << mapping.PrintCompact()
-                      << std::endl;
-          mutex_->unlock();
-        }
-
-        mappings_since_last_best_update = 0;
-      }
-      else
-      {
-        // If the only change in this mapping vs. the previous mapping was in
-        // its dataspace bypass scheme, then we may not want to make this
-        // failure count towards the timeout termination trigger.
-        if (penalize_consecutive_bypass_fails_ || !only_bypass_changed)
-        {
-          mappings_since_last_best_update++;
-        }
-      }
-    } // while ()
-      
-    //
-    // End Mapping.
-    //
-  }
 };
